+++ doc/Changelog.md
@@ -2,6 +2,9 @@
 
 # git master
 
+* [67](https://github.com/HBPVIS/ZeroBuf/pull/67):
+  * Fix compilation error when having multiple FBS enums in one namespace
+  * Maintain order of FBS fields in JSON schema
 * [65](https://github.com/HBPVIS/ZeroBuf/pull/65):
   * JSON schema generated from FBS, available in ZeroBuf::getSchema() and ZeroBuf::ZEROBUF_SCHEMA
   * enum class for enums instead of 'old' enums; breaks existing enum names